% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicom_header.R
\name{all_valid_header_names}
\alias{all_valid_header_names}
\title{Get all valid DICOM header names}
\usage{
all_valid_header_names()
}
\value{
Vector of all possible header keywords (e.g. "Patient's Name") from the DICOM standard
}
\description{
Get all valid DICOM header names
}
