% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicom_header.R
\name{dicom_all_valid_header_tags}
\alias{dicom_all_valid_header_tags}
\title{Get all valid DICOM header tags}
\usage{
dicom_all_valid_header_tags()
}
\value{
Vector of all possible header tags (e.g. "(0008,0020)") from the DICOM standard
}
\description{
Get all valid DICOM header tags
}
