% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_dicom.R
\name{dicom_constant_header_values}
\alias{dicom_constant_header_values}
\title{Get the values of header attributes that are constant across slices}
\usage{
dicom_constant_header_values(dicom_data, numeric = TRUE)
}
\arguments{
\item{dicom_data}{DICOM data returned by \code{\link{read_dicom}}}

\item{numeric}{Convert number values to numeric instead of strings}
}
\value{
List of field values that are constant across all slices. List identifiers
are field names and values are the common attribute values. Fields that are included
more than once in the header are excluded from the return list.
}
\description{
Get the values of header attributes that are constant across slices
}
