% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_import.R
\name{dicom_data_to_mat}
\alias{dicom_data_to_mat}
\title{Convert DICOM data to 3D matrix of intensities}
\usage{
dicom_data_to_mat(dicom_data)
}
\arguments{
\item{dicom_data}{DICOM data returned by \code{\link{read_dicom}}}
}
\value{
3D array of intensities
}
\description{
Convert DICOM data to 3D matrix of intensities
}
