% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_dicom.R
\name{dicom_header_tag}
\alias{dicom_header_tag}
\title{Get DICOM header tag string corresponding to a group and element}
\usage{
dicom_header_tag(group, element)
}
\arguments{
\item{group}{Group e.g. "0008"}

\item{element}{Element e.g. "0020"}
}
\value{
The tag e.g. "(0008,0020)"
}
\description{
Get DICOM header tag string corresponding to a group and element
}
\examples{
dicom_header_tag("0008", "0020")
}
