% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicom_header.R
\name{dicom_header_values}
\alias{dicom_header_values}
\title{Get vector of header values for each DICOM slice for a particular header field}
\usage{
dicom_header_values(dicom_data, field, numeric = TRUE)
}
\arguments{
\item{dicom_data}{DICOM data returned by \code{\link{read_dicom}}}

\item{field}{Header field keyword e.g. "PatientName"}

\item{numeric}{Convert values to numbers}
}
\value{
Vector of header values
}
\description{
Get vector of header values for each DICOM slice for a particular header field
}
