% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_dicom.R
\name{dicom_search_header_keywords}
\alias{dicom_search_header_keywords}
\title{Search header keywords in the DICOM standard for matches to a string}
\usage{
dicom_search_header_keywords(str)
}
\arguments{
\item{str}{String to search for (case insensitive)}
}
\value{
Vector of header keywords (e.g. "PatientName") matching the string
}
\description{
Search header keywords in the DICOM standard for matches to a string
}
