% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_dicom.R
\name{dicom_search_header_names}
\alias{dicom_search_header_names}
\title{Search header names in the DICOM standard for matches to a string}
\usage{
dicom_search_header_names(str)
}
\arguments{
\item{str}{String to search for (case insensitive)}
}
\value{
Vector of header names (e.g. "Patient's Name") matching the string
}
\description{
Search header names in the DICOM standard for matches to a string
}
