% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_dicom.R
\name{dicom_standard_timestamp}
\alias{dicom_standard_timestamp}
\title{Get the time at which the DICOM standard was loaded from the web for this package}
\usage{
dicom_standard_timestamp()
}
\value{
Timestamp
}
\description{
Get the time at which the DICOM standard was loaded from the web for this package
}
