% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicom_header.R
\name{dicom_standard_version}
\alias{dicom_standard_version}
\title{Get the version of the DICOM standard assumed by validation functions}
\usage{
dicom_standard_version()
}
\value{
DICOM standard version
}
\description{
Get the version of the DICOM standard assumed by validation functions
}
