% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicom_header.R
\name{dicom_standard_web}
\alias{dicom_standard_web}
\title{Get the website used to load the DICOM standard for this package}
\usage{
dicom_standard_web()
}
\value{
Web URL for DICOM standard
}
\description{
Get the website used to load the DICOM standard for this package
}
