% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicom_header.R
\name{dicom_validate_has_field}
\alias{dicom_validate_has_field}
\title{Check that a field exists in DICOM header}
\usage{
dicom_validate_has_field(dicom_data, field)
}
\arguments{
\item{dicom_data}{DICOM data}

\item{field}{Field name}
}
\description{
Check that a field exists in DICOM header
}
