% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_dicom.R
\name{dicom_validate_keyword}
\alias{dicom_validate_keyword}
\title{Validate a header keyword against the DICOM standard}
\usage{
dicom_validate_keyword(keyword, stop = TRUE)
}
\arguments{
\item{keyword}{Keyword}

\item{stop}{If true, raise error when validation fails. If false, raise warning.}
}
\description{
Validate a header keyword against the DICOM standard
}
\keyword{internal}
