% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_dicom.R
\name{dicom_validate_tag}
\alias{dicom_validate_tag}
\title{Validate a header tag against the DICOM standard}
\usage{
dicom_validate_tag(tag, stop = TRUE)
}
\arguments{
\item{tag}{Tag}

\item{stop}{If true, raise error when validation fails. If false, raise warning.}
}
\description{
Validate a header tag against the DICOM standard
}
\keyword{internal}
