% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicom_header.R
\name{header_as_matrix}
\alias{header_as_matrix}
\title{Get the header information as a matrix}
\usage{
header_as_matrix(dicom_data, slice_idx = NA)
}
\arguments{
\item{dicom_data}{DICOM data returned by \code{\link{read_dicom}}}

\item{slice_idx}{1-based slice index. If NA, all slices will be included.}
}
\value{
Data frame containing one record for each header attribute. Note that
if all slices are included, fields that appear more than once in a given
slice header will be excluded from the values reported for that slice.
}
\description{
Get the header information as a matrix
}
