% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_dicom.R
\name{header_fields.dicomdata}
\alias{header_fields.dicomdata}
\title{Get the names of DICOM header fields for an image series.
If field names are repeated within a single header, these duplicate
fields are omitted from the return value. If slices have different
header fields, this function returns the union across slices of
all field names.}
\usage{
\method{header_fields}{dicomdata}(img_data)
}
\arguments{
\item{img_data}{DICOM data returned by \code{\link{read_dicom}}}
}
\value{
Vector of header field names
}
\description{
Get the names of DICOM header fields for an image series.
If field names are repeated within a single header, these duplicate
fields are omitted from the return value. If slices have different
header fields, this function returns the union across slices of
all field names.
}
