% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_nifti.R
\name{header_fields.nifti1data}
\alias{header_fields.nifti1data}
\title{Get the fields in a NIfTI-1 header.}
\usage{
\method{header_fields}{nifti1data}(img_data)
}
\arguments{
\item{img_data}{NIfTI-1 data returned by \code{\link{read_nifti1}}}
}
\value{
Vector of header field names
}
\description{
See the \href{https://nifti.nimh.nih.gov/pub/dist/src/niftilib/nifti1.h}{official definition of the NIfTI-1 header}.
}
