% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{header_fields}
\alias{header_fields}
\title{Get the names of metadata fields from an image dataset's header(s)}
\usage{
header_fields(img_data)
}
\arguments{
\item{img_data}{Image data returned by e.g. \code{\link{read_dicom}} or \code{\link{read_nifti1}}}
}
\value{
Vector of header field names
}
\description{
Get the names of metadata fields from an image dataset's header(s)
}
\examples{
\dontrun{header_fields(dicom_data)}
\dontrun{header_fields(nifti1_data)}
}
