% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicom_header.R
\name{header_tag}
\alias{header_tag}
\title{Get DICOM header tag string corresponding to a group and element}
\usage{
header_tag(group, element)
}
\arguments{
\item{group}{Group e.g. "0008"}

\item{element}{Element e.g. "0020"}
}
\value{
The tag e.g. "(0008,0020)"
}
\description{
Get DICOM header tag string corresponding to a group and element
}
