% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_dicom.R
\name{header_value.dicomdata}
\alias{header_value.dicomdata}
\title{Get vector of header values for each DICOM slice for a header field}
\usage{
\method{header_value}{dicomdata}(img_data, field)
}
\arguments{
\item{img_data}{DICOM data returned by \code{\link{read_dicom}}}

\item{field}{Header field keyword e.g. "PatientName"}
}
\value{
Vector of header values. Numeric values are converted to numbers.
}
\description{
Get vector of header values for each DICOM slice for a header field
}
