% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{header_value}
\alias{header_value}
\title{Get metadata contained in a header field}
\usage{
header_value(img_data, field)
}
\arguments{
\item{img_data}{Image data returned by e.g. \code{\link{read_dicom}} or \code{\link{read_nifti1}}}

\item{field}{Field name}
}
\value{
Metadata for the field in an appropriate format for the data type
}
\description{
Get metadata contained in a header field
}
\examples{
data(sample_dicom_img)
data(sample_nifti_img)
header_value(sample_dicom_img, "SliceLocation")
header_value(sample_nifti_img, "dim_")
}
