% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_import.R
\name{img_data_to_mat}
\alias{img_data_to_mat}
\title{Convert image data to 3D matrix of intensities}
\usage{
img_data_to_mat(img_data)
}
\arguments{
\item{img_data}{Image data returned by e.g. \code{\link{read_dicom}} or \code{\link{read_nifti}}}
}
\value{
3D array of intensities where third dimension is slice
}
\description{
Convert image data to 3D matrix of intensities
}
