% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{img_dimensions}
\alias{img_dimensions}
\title{Get the dimensions of an image}
\usage{
img_dimensions(img_data)
}
\arguments{
\item{img_data}{Image data returned by e.g. \code{\link{read_dicom}} or \code{\link{read_nifti1}}}
}
\value{
Image dimensions
}
\description{
Get the dimensions of an image
}
\examples{
\dontrun{img_dimensions(dicom_data)}
\dontrun{img_dimensions(nifti1_data)}
}
