% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_import.R
\name{mat_reduce_dim}
\alias{mat_reduce_dim}
\title{Take a slice of a multidimensional array by fixing the coordinates in the last dimension(s)}
\usage{
mat_reduce_dim(mat, coords_last_dims)
}
\arguments{
\item{mat}{Multidimensional array}

\item{coords_last_dims}{One or more coordinates in the last one or more dimensions of the matrix}
}
\value{
The reduced matrix with last dimension(s) collapsed
}
\description{
Take a slice of a multidimensional array by fixing the coordinates in the last dimension(s)
}
\keyword{internal}
