% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_nifti.R
\name{nifti1_header_values}
\alias{nifti1_header_values}
\title{Get named list of header attributes for a NIfTI-1 file}
\usage{
nifti1_header_values(img_data)
}
\arguments{
\item{img_data}{NIfTI-1 data returned by \code{\link{read_nifti1}}}
}
\value{
List of header attribute values
}
\description{
Get named list of header attributes for a NIfTI-1 file
}
\examples{
data(sample_nifti_img)
nifti1_header_values(sample_nifti_img)
}
