% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_nifti.R
\name{nifti1_num_dim}
\alias{nifti1_num_dim}
\title{Returns the number of dimensions in a NIfTI-1 image}
\usage{
nifti1_num_dim(nifti1_data)
}
\arguments{
\item{nifti1_data}{NIfTI-1 data returned by \code{\link{read_nifti1}}}
}
\description{
Returns the number of dimensions in a NIfTI-1 image
}
