% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicom_header.R
\name{num_slices_dicom}
\alias{num_slices_dicom}
\title{Get the number of image slices in a DICOM series}
\usage{
num_slices_dicom(dicom_data)
}
\arguments{
\item{dicom_data}{DICOM data returned by \code{\link{read_dicom}}}
}
\value{
Number of slices
}
\description{
Get the number of image slices in a DICOM series
}
