% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_import.R
\name{read_dicom}
\alias{read_dicom}
\title{Read a series of DICOM images from a directory}
\usage{
read_dicom(dir)
}
\arguments{
\item{dir}{Directory containing DICOM images}
}
\value{
List with elements \code{hdr} and \code{img}, each with an element for each slice
}
\description{
Read a series of DICOM images from a directory
}
