% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_dicom_img}
\alias{sample_dicom_img}
\title{A sample DICOM image}
\format{A `dicomdata` object; list with elements `hdr` and `img`
\describe{
  \item{hdr}{DICOM header}
  \item{img}{DICOM image data}
}}
\source{
\url{http://www.pcir.org/researchers/downloads_available.html}
}
\usage{
sample_dicom_img
}
\description{
A sample DICOM image
}
\keyword{datasets}
