% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_nifti_img}
\alias{sample_nifti_img}
\title{A sample NIfTI-1 image}
\format{List with one element `data`
\describe{
  \item{data}{Object of class `nifti`}
}}
\source{
\url{https://nifti.nimh.nih.gov/nifti-1/data}
}
\usage{
sample_nifti_img
}
\description{
A sample NIfTI-1 image
}
\keyword{datasets}
