% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicom_header.R
\name{validate_group_element}
\alias{validate_group_element}
\title{Validate a header group and element against the DICOM standard}
\usage{
validate_group_element(group, element, stop = TRUE)
}
\arguments{
\item{group}{Group}

\item{element}{Element}

\item{stop}{If true, raise error when validation fails. If false, raise warning.}
}
\description{
Validate a header group and element against the DICOM standard
}
