% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicom_header.R
\name{validate_keyword}
\alias{validate_keyword}
\title{Validate a header keyword against the DICOM standard}
\usage{
validate_keyword(keyword, stop = TRUE)
}
\arguments{
\item{keyword}{Keyword}

\item{stop}{If true, raise error when validation fails. If false, raise warning.}
}
\description{
Validate a header keyword against the DICOM standard
}
