% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_dicom.R
\name{validate_metadata.dicomdata}
\alias{validate_metadata.dicomdata}
\title{Validate all header elements (group, element, name) for all slices}
\usage{
\method{validate_metadata}{dicomdata}(dicom_data, stop = TRUE)
}
\arguments{
\item{dicom_data}{DICOM data returned by \code{\link{read_dicom}}}

\item{stop}{If true, raise error when validation fails. If false, raise warnings for each failed element.}
}
\description{
Validate all header elements (group, element, name) for all slices
}
