% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{validate_metadata}
\alias{validate_metadata}
\title{Validate the metadata associated with an image dataset}
\usage{
validate_metadata(img_data, stop = TRUE)
}
\arguments{
\item{img_data}{Image data returned by e.g. \code{\link{read_dicom}} or \code{\link{read_nifti}}}

\item{stop}{If true, raise error when validation fails. If false, raise warnings for each failed element.}
}
\description{
Validate the metadata associated with an image dataset
}
