% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicom_header.R
\name{validate_tag}
\alias{validate_tag}
\title{Validate a header tag against the DICOM standard}
\usage{
validate_tag(tag, stop = TRUE)
}
\arguments{
\item{tag}{Tag}

\item{stop}{If true, raise error when validation fails. If false, raise warning.}
}
\description{
Validate a header tag against the DICOM standard
}
