% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view_slice_mat}
\alias{view_slice_mat}
\title{Display a visual of one slice of an image matrix}
\usage{
view_slice_mat(mat, slice = NULL, col = grey(0:64/64), ...)
}
\arguments{
\item{slice}{Slice number, or NULL if matrix is 2D}

\item{col}{Color scheme}

\item{img_data}{2D or 3D intensity matrix, e.g. the return value from \code{\link{img_data_to_mat}} or \code{\link{img_data_to_3D_mat}}}
}
\description{
Display a visual of one slice of an image matrix
}
