% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicom_header.R
\name{header_fields}
\alias{header_fields}
\title{Get the names of DICOM header fields for an image series.
If field names are repeated within a single header, these duplicate
fields are omitted from the return value. If slices have different
header fields, this function returns the union across slices of
all field names.}
\usage{
header_fields(dicom_data)
}
\arguments{
\item{dicom_data}{DICOM data returned by \code{\link{read_dicom}}}
}
\value{
Vector of header field names
}
\description{
Get the names of DICOM header fields for an image series.
If field names are repeated within a single header, these duplicate
fields are omitted from the return value. If slices have different
header fields, this function returns the union across slices of
all field names.
}
