% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_import.R
\name{read_dicom}
\alias{read_dicom}
\title{Read a DICOM image or series of images}
\usage{
read_dicom(path, ...)
}
\arguments{
\item{path}{Directory containing DICOM images, or single image file}

\item{...}{Additional arguments to \code{\link[oro.dicom]{readDICOM}}}
}
\value{
List with elements \code{hdr} and \code{img}, each with an element for each slice
}
\description{
Read a DICOM image or series of images
}
\examples{
\dontrun{read_dicom(dicom_directory)}
\dontrun{read_dicom(dicom_file.dcm)}
}
