% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_import.R
\name{read_nifti1}
\alias{read_nifti1}
\title{Read a NIfTI-1 image}
\usage{
read_nifti1(file, ...)
}
\arguments{
\item{file}{.nii file, gzipped or not, or base of .hdr and .img files without extension}

\item{...}{Additional arguments to \code{\link[oro.nifti]{readNIfTI}}}
}
\value{
List containing object of class \code{\link[oro.nifti]{nifti}}
}
\description{
Read a NIfTI-1 image
}
\examples{
\dontrun{read_nifti1(nifti_file.nii)}
\dontrun{read_nifti1(nifti_file_basename)}
}
